add office31 domain adaptation to examples file
